Add signup components for desktop and mobile
